requirements.txt
beautifulsoup4==4.6.0
certifi==2018.1.18
chardet==3.0.4
future==0.16.0
idna==2.6
jieba==0.39
multidict==4.1.0
Pillow==5.0.0
pytesseract==0.1.8
PyYAML==3.12
requests==2.18.4
selenium==2.53.6
six==1.11.0
Unidecode==1.0.22
urllib3==1.22
vcrpy==1.11.1
wrapt==1.10.11
yarl==1.1.0
